Test files are classified.
