+++ tests/analyzer/test_duplicate.py
@@ -1,228 +1,7 @@
-import textwrap
 import unittest
 
-from unimport.analyzer import Analyzer
+from tests.analyzer.utils import UnusedTestCase
 from unimport.statement import Import, ImportFrom
-
-
-class UnusedTestCase(unittest.TestCase):
-    maxDiff = None
-    include_star_import = False
-
-    def assertSourceAfterScanningEqualToExpected(
-        self, source, expected_unused_imports=[]
-    ):
-        analyzer = Analyzer(
-            source=textwrap.dedent(source),
-            include_star_import=self.include_star_import,
-        )
-        analyzer.traverse()
-        super().assertEqual(
-            expected_unused_imports,
-            list(analyzer.get_unused_imports()),
-        )
-        analyzer.clear()
-
-
-class BuiltinTestCase(UnusedTestCase):
-    # https://github.com/hakancelik96/unimport/issues/45
-
-    def test_ConnectionError(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            from x.y import ConnectionError
-            try:
-               pass
-            except ConnectionError:
-               pass
-            """
-        )
-
-    def test_ValueError(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            from x import ValueError
-            print(ValueError)
-            """
-        )
-
-    def test_builtins(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            from builtins import next, object, range
-            __all__ = ["next", "object"]
-            for i in range(8):
-               pass
-            """
-        )
-
-
-class AllTestCase(UnusedTestCase):
-    def test_from_import(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            from codeop import compile_command
-            __all__ = ["compile_command"]
-            """
-        )
-
-    def test_defined_top(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            source="""\
-            __all__ = ["x"]
-            import x
-            """,
-        )
-
-
-class UnusedImportTestCase(UnusedTestCase):
-    def test_comma(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            from os import (
-                waitpid,
-                scandir,
-            )
-            """,
-            [
-                ImportFrom(
-                    lineno=1,
-                    column=1,
-                    suggestions=[],
-                    name="waitpid",
-                    package="os",
-                    star=False,
-                ),
-                ImportFrom(
-                    lineno=1,
-                    column=2,
-                    suggestions=[],
-                    name="scandir",
-                    package="os",
-                    star=False,
-                ),
-            ],
-        )
-
-    def test_module_used(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            from pathlib import Path
-            CURRENT_DIR = Path(".").parent
-            """
-        )
-
-    def test_module_unused(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            from pathlib import Path
-            """,
-            [
-                ImportFrom(
-                    lineno=1,
-                    column=1,
-                    suggestions=[],
-                    name="Path",
-                    package="pathlib",
-                    star=False,
-                ),
-            ],
-        )
-
-    def test_unknown_module_used(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            import x.y
-            CURRENT_DIR = x.y(".").parent
-            """
-        )
-
-    def test_unknown_module_unused(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            import x.y
-            import d.f.a.s
-            CURRENT_DIR = x.y(".").parent
-            """,
-            [
-                Import(
-                    lineno=2,
-                    column=1,
-                    name="d.f.a.s",
-                    package="d.f.a.s",
-                ),
-            ],
-        )
-
-    def test_import_in_function(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            import t
-            from x import y, z
-
-            def function(f=t):
-                import x
-                return f
-            from i import t, ii
-            print(t)
-            """,
-            [
-                ImportFrom(
-                    lineno=2,
-                    column=1,
-                    name="y",
-                    package="x",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=2,
-                    column=2,
-                    name="z",
-                    package="x",
-                    star=False,
-                    suggestions=[],
-                ),
-                Import(
-                    lineno=5,
-                    column=1,
-                    name="x",
-                    package="x",
-                ),
-                ImportFrom(
-                    lineno=7,
-                    column=2,
-                    name="ii",
-                    package="i",
-                    star=False,
-                    suggestions=[],
-                ),
-            ],
-        )
-
-    def test_import_after_usage(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            def function():
-                print(os)
-                import os
-            """,
-            [Import(lineno=3, column=1, name="os", package="os")],
-        )
-
-    def test_double_underscore_builtins_names(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            from globals import (
-                __name__, __doc__, __package__,
-               __loader__, __spec__, __annotations__,
-               __builtins__
-            )
-            __name__, __doc__, __package__,
-            __loader__, __spec__, __annotations__,
-            __builtins__
-            """
-        )
 
 
 @unittest.skip("Temporarily removed.")
@@ -1012,195 +791,4 @@ class DuplicateTestCase(UnusedTestCase):
                     suggestions=[],
                 ),
             ],
-        )
-
-
-class AsImportTestCase(UnusedTestCase):
-    def test_as_import_all_unused_all_cases(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            from x import y as z
-            import x
-            from t import s as ss
-            from f import a as c, l as k, i as ii
-            from fo import (bar, i, x as z)
-            import le as x
-            """,
-            [
-                ImportFrom(
-                    lineno=1,
-                    column=1,
-                    name="z",
-                    package="x",
-                    star=False,
-                    suggestions=[],
-                ),
-                Import(
-                    lineno=2,
-                    column=1,
-                    name="x",
-                    package="x",
-                ),
-                ImportFrom(
-                    lineno=3,
-                    column=1,
-                    name="ss",
-                    package="t",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=4,
-                    column=1,
-                    name="c",
-                    package="f",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=4,
-                    column=2,
-                    name="k",
-                    package="f",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=4,
-                    column=3,
-                    name="ii",
-                    package="f",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=5,
-                    column=1,
-                    name="bar",
-                    package="fo",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=5,
-                    column=2,
-                    name="i",
-                    package="fo",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=5,
-                    column=3,
-                    name="z",
-                    package="fo",
-                    star=False,
-                    suggestions=[],
-                ),
-                Import(
-                    lineno=6,
-                    column=1,
-                    name="x",
-                    package="le",
-                ),
-            ],
-        )
-
-    @unittest.skip("Temporarily removed.")
-    def test_as_import_one_used_in_function_all_cases(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            from x import y as z
-            import x
-            from t import s as ss
-            from f import a as c, l as k, i as ii
-            from fo import (bar, i, x as z)
-            import le as x
-            def x(t=x):pass
-            """,
-            [
-                ImportFrom(
-                    lineno=1,
-                    column=1,
-                    name="z",
-                    package="x",
-                    star=False,
-                    suggestions=[],
-                ),
-                Import(
-                    lineno=2,
-                    column=1,
-                    name="x",
-                    package="x",
-                ),
-                ImportFrom(
-                    lineno=3,
-                    column=1,
-                    name="ss",
-                    package="t",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=4,
-                    column=1,
-                    name="c",
-                    package="f",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=4,
-                    column=2,
-                    name="k",
-                    package="f",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=4,
-                    column=3,
-                    name="ii",
-                    package="f",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=5,
-                    column=1,
-                    name="bar",
-                    package="fo",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=5,
-                    column=2,
-                    name="i",
-                    package="fo",
-                    star=False,
-                    suggestions=[],
-                ),
-                ImportFrom(
-                    lineno=5,
-                    column=3,
-                    name="z",
-                    package="fo",
-                    star=False,
-                    suggestions=[],
-                ),
-            ],
-        )
-
-
-class DealingImplicitImportsSubPackagesTestCase(UnusedTestCase):
-    # https://github.com/hakancelik96/unimport/issues/127
-
-    def test(self):
-        self.assertSourceAfterScanningEqualToExpected(
-            """\
-            import x.y
-
-            x
-            """,
         )
