+++ tests/analyzer/test_implicit_import.py
@@ -1,0 +1,14 @@
+from tests.analyzer.utils import UnusedTestCase
+
+
+class DealingImplicitImportsSubPackagesTestCase(UnusedTestCase):
+    # https://github.com/hakancelik96/unimport/issues/127
+
+    def test(self):
+        self.assertSourceAfterScanningEqualToExpected(
+            """\
+            import x.y
+
+            x
+            """,
+        )
